+++ lib/braintree/paypal_account.rb
@@ -7,6 +7,7 @@ module Braintree
     attr_reader :customer_id
     attr_reader :email
     attr_reader :funding_source_description
+    attr_reader :edit_paypal_vault_id
     attr_reader :image_url
     attr_reader :payer_id
     attr_reader :revoked_at